// apv-angular/src/app/core/models/Donation.ts
import { DonationParticipant } from "./DonationParticipant";

export interface Donation {
  id?: number;

  publicId?: string;

  title: string;

  description: string;

  dateDonation: Date;

  closure_status?: boolean;

  beneficiaryId: number|null;

  beneficiaryName?: string;

  photo?: string;

  publicStatus?: boolean;

  participants?: DonationParticipant[];
}